+++ packages/sdk/dappkit/src/index.ts
@@ -20,36 +20,6 @@ export {
   SignTxRequest,
 } from '@celo/utils'
 
-// DappKit Web constants and helpers
-const localStorageKey = 'dappkit-web'
-// hack to get around dappkit issue where new tabs are opened
-// and the url hash state is not respected (Note this implementation
-// of dappkit doesn't use URL hashes to always force the newtab experience).
-
-// Function that should be called within dapp wherever a window is rendered
-export function parseURLOnRender() {
-  if (typeof window !== 'undefined') {
-    const params = new URL(window.location.href).searchParams
-    if (params.get('type') && params.get('requestId')) {
-      localStorage.setItem(localStorageKey, window.location.href)
-      // TODO: seems like the below line is not getting executed in Chrome on iOS
-      // on iOS though, the newly opened window gets closed and the following window (instead of the previously opened tab) gets opened
-      window.close()
-    }
-  }
-}
-
-async function waitForResponse() {
-  while (true) {
-    const value = localStorage.getItem(localStorageKey)
-    if (value) {
-      localStorage.removeItem(localStorageKey)
-      return value
-    }
-    await new Promise((resolve) => setTimeout(resolve, 100))
-  }
-}
-
 export function listenToAccount(callback: (account: string) => void) {
   return Linking.addEventListener('url', ({ url }: { url: string }) => {
     try {
@@ -78,7 +48,7 @@ export function listenToSignedTxs(callback: (signedTxs: string[]) => void) {
   })
 }
 
-function waitDecoratorMobile(
+function waitDecorator(
   requestId: string,
   checkCallback: (requestId: string, dappKitResponse: any) => boolean
 ): Promise<any> {
@@ -98,20 +68,7 @@ function waitDecoratorMobile(
   })
 }
 
-async function waitDecoratorWeb(
-  requestId: string,
-  checkCallback: (requestId: string, dappKitResponse: any) => boolean
-): Promise<any> {
-  const url = await waitForResponse()
-  const dappKitResponse = parseDappkitResponseDeeplink(url)
-  if (checkCallback(requestId, dappKitResponse)) {
-    return dappKitResponse
-  }
-  console.log('Unable to parse url', url)
-  throw new Error('Unable to parse Valora response')
-}
-
-function checkAccountAuth(requestId: string, dappKitResponse: any): boolean {
+export function checkAccountAuth(requestId: string, dappKitResponse: any): boolean {
   return (
     requestId === dappKitResponse.requestId &&
     dappKitResponse.type === DappKitRequestTypes.ACCOUNT_ADDRESS &&
@@ -119,7 +76,7 @@ function checkAccountAuth(requestId: string, dappKitResponse: any): boolean {
   )
 }
 
-function checkSignedTxs(requestId: string, dappKitResponse: any): boolean {
+export function checkSignedTxs(requestId: string, dappKitResponse: any): boolean {
   return (
     requestId === dappKitResponse.requestId &&
     dappKitResponse.type === DappKitRequestTypes.SIGN_TX &&
@@ -127,24 +84,12 @@ function checkSignedTxs(requestId: string, dappKitResponse: any): boolean {
   )
 }
 
-// For usage in Mobile DApps
 export function waitForAccountAuth(requestId: string): Promise<AccountAuthResponseSuccess> {
-  return waitDecoratorMobile(requestId, checkAccountAuth)
+  return waitDecorator(requestId, checkAccountAuth)
 }
 
 export function waitForSignedTxs(requestId: string): Promise<SignTxResponseSuccess> {
-  return waitDecoratorMobile(requestId, checkSignedTxs)
-}
-
-// For usage in Web DApps
-export async function waitForAccountAuthWeb(
-  requestId: string
-): Promise<AccountAuthResponseSuccess> {
-  return waitDecoratorWeb(requestId, checkAccountAuth)
-}
-
-export async function waitForSignedTxsWeb(requestId: string): Promise<SignTxResponseSuccess> {
-  return waitDecoratorWeb(requestId, checkSignedTxs)
+  return waitDecorator(requestId, checkSignedTxs)
 }
 
 export function requestAccountAddress(meta: DappKitRequestMeta) {
